[SDLPointer] adds 'pass(to:)' & 'result(of:)'
`SDLPointer` is a Swift class that can be used to pass its pointer
to any SDL function call, currently with up to four arguments.

Use `pass(to:)` if the SDL function call is not capable of causing
an error.

Use `result(of:)` if the SDL function call can return a value which
indicates an error may have occurred.

My goal in these changes is to minimize the need to explicitly wrap
every possible SDL call with Swift code, and instead use instances
of `SDLPointer` that manage an underlying pointer to SDL objects.